adding Improvements to master branch

+++ app/Http/Controllers/AuthController.php
@@ -9,40 +9,68 @@ use App\Models\User;
 use App\Http\Controllers\BaseController;
 use Symfony\Component\HttpFoundation\Response;
 use Config;
+use App\Http\Resources\AuthResource;
+use App\Exceptions;
 
 class AuthController extends BaseController
 {
 
     public function login(Request $request) {
-        $this->validateHeaders($request);
-        $validator = Validator::make($request->all(), [
-            "email" =>  "required|email",
-            "password" =>  "required",
-        ]);
+        try {
+            $requestHeaders = $this->validateHeaders($request);
+            if(!$requestHeaders['isValid'])
+            {
+                return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+            }
 
-        if($validator->fails())
-            return $this->sendError($validator->errors());
+            $validator = Validator::make($request->all(), [
+                'email' =>  'required|email',
+                'password' =>  'required',
+            ]);
 
-        $user = User::where("email", $request->email)->first();
+            if($validator->fails())
+            {
+                return $this->sendError($validator->errors());
+            }
 
-        if(is_null($user))
-            return $this->sendError([Config::get('constants.messages.email_not_found')]);
+            $user = User::where('email', $request->email)->first();
 
-        if(!Auth::attempt($request->only('email', 'password')))
-            return $this->sendError([Config::get('constants.messages.invalid_password')], Response::HTTP_UNAUTHORIZED);
+            if(is_null($user))
+            {
+                return $this->sendError([Config::get('constants.messages.email_not_found')]);
+            }
+
+            if(!Auth::attempt($request->only('email', 'password')))
+            {
+                return $this->sendError([Config::get('constants.messages.invalid_password')], Response::HTTP_UNAUTHORIZED);
+            }
+            
+            $tokenResult = User::where('email', $request->email)
+                            ->first()
+                            ->createToken('authToken')
+                            ->plainTextToken;
+            
+            return $this->sendResponse(['token' => $tokenResult]);
+        } catch (Exceptions $e) {
+            return $this->sendError($e->getMessage());
+        }
         
-        $tokenResult = User::where("email", $request->email)
-                        ->first()
-                        ->createToken('authToken')
-                        ->plainTextToken;
-        
-        return $this->sendResponse($tokenResult);
     }
 
     public function logout(Request $request)
     {
-        $this->validateHeaders($request);
-        $request->user()->currentAccessToken()->delete();
-        return $this->sendResponse(null);
+        try {
+            $requestHeaders = $this->validateHeaders($request);
+            if(!$requestHeaders['isValid'])
+            {
+                return $this->sendError($requestHeaders['message'],$requestHeaders['code']);
+            }
+            
+            $request->user()->currentAccessToken()->delete();
+            return $this->sendResponse(null);
+        } catch (Exceptions $e) {
+            return $this->sendError($e->getMessage());
+        }
+        
     }
 }
